Move foreign key test into pgslice_test.rb
