+++ test/pgslice_test.rb
@@ -1,21 +1,21 @@
 require_relative "test_helper"
 
 class PgSliceTest < Minitest::Test
-  def setup
-    create_tables
-    populate_data
-  end
-
-  def teardown
-    drop_tables
-  end
-
   def test_day
     assert_period("day")
   end
 
   def test_month
     assert_period("month")
+  end
+
+  def test_foreign_keys
+    assert has_foreign_key?("Posts"), "Original table is missing foreign key"
+    run_command "prep Posts createdAt month"
+    assert has_foreign_key?("Posts_intermediate"), "Intermediate table is missing foreign key"
+    run_command "add_partitions Posts --intermediate --past 1 --future 1"
+    assert has_foreign_key?("Posts_#{Time.now.strftime("%Y%m")}"), "Partition is missing foreign key"
+    run_command "unprep Posts"
   end
 
   private
@@ -38,5 +38,15 @@ class PgSliceTest < Minitest::Test
     puts
     PgSlice::Client.new("#{command} --url pgslice_test".split(" ")).perform
     puts
+  end
+
+  def has_foreign_key?(table_name)
+    @conn ||= PG::Connection.open(dbname: "pgslice_test")
+    result = @conn.exec <<-SQL
+      SELECT pg_get_constraintdef(oid) AS def
+      FROM pg_constraint
+      WHERE contype = 'f' AND conrelid = '"#{table_name}"'::regclass
+    SQL
+    !!result.detect { |row| row['def'] =~ /\AFOREIGN KEY \(.*\) REFERENCES "Users"\("Id"\)\z/ }
   end
 end
